# docs/api/swagger/api-v1-users/server.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api-v1-users/server.proto",
    "version": "version not set"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/users/auth": {
      "post": {
        "operationId": "Auth",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/usersAuthResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/gatewayruntimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/usersAuthRequest"
            }
          }
        ],
        "tags": [
          "UsersAPIService"
        ]
      }
    },
    "/v1/users/user": {
      "post": {
        "operationId": "CreateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/usersCreateUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/gatewayruntimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/usersCreateUserRequest"
            }
          }
        ],
        "tags": [
          "UsersAPIService"
        ]
      }
    },
    "/v1/users/user/{uuid}/role": {
      "post": {
        "operationId": "AddRole",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/gatewayruntimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/usersAddRoleRequest"
            }
          }
        ],
        "tags": [
          "UsersAPIService"
        ]
      }
    }
  },
  "definitions": {
    "gatewayruntimeError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "usersAddRoleRequest": {
      "type": "object",
      "properties": {
        "uuid": {
          "type": "string"
        },
        "role": {
          "$ref": "#/definitions/usersRole"
        }
      }
    },
    "usersAuthRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "usersAuthResponse": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "valid": {
          "type": "boolean",
          "format": "boolean"
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1usersError"
          }
        }
      }
    },
    "usersCreateUserRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "company": {
          "type": "string"
        }
      }
    },
    "usersCreateUserResponse": {
      "type": "object",
      "properties": {
        "stateCode": {
          "$ref": "#/definitions/usersResponseStateCode"
        }
      }
    },
    "usersResponseStateCode": {
      "type": "string",
      "enum": [
        "CREATED"
      ],
      "default": "CREATED"
    },
    "usersRole": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "USER",
        "USER_ADMIN"
      ],
      "default": "UNKNOWN"
    },
    "v1usersError": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "description": {
          "type": "string"
        }
      }
    }
  }
}